+++ sdk/log/record.go
@@ -120,6 +120,7 @@ type Record struct {
 	scope *instrumentation.Scope
 
 	attributeValueLengthLimit int
+	attributeValueDepthLimit  int
 	attributeCountLimit       int
 
 	// allowDupKeys specifies whether duplicate keys are allowed in key-value
@@ -279,6 +280,7 @@ func (r *Record) AddAttributes(attrs ...attribute.KeyValue) {
 				// New attrs overwrite any existing with the same key.
 				dropped++
 				logKeyValuePairDropped()
+				a = r.applyAttrLimitsAndDedup(a)
 				if idx < 0 {
 					r.front[-(idx + 1)] = a
 				} else {
@@ -514,13 +516,35 @@ func (r *Record) Clone() Record {
 }
 
 func (r *Record) applyAttrLimitsAndDedup(attr attribute.KeyValue) attribute.KeyValue {
-	if !r.allowDupKeys {
-		var changed bool
-		attr, changed = attrnorm.KeyValue(attr)
-		if changed {
-			logKeyValuePairDropped()
-		}
+	switch attr.Value.Type() {
+	case attribute.SLICE, attribute.MAP:
+		attr = r.applyCompositeAttrDepthLimitAndDedup(attr)
 	}
 	attr.Value = attrnorm.TruncateValue(r.attributeValueLengthLimit, attr.Value)
 	return attr
+}
+
+func (r *Record) applyCompositeAttrDepthLimitAndDedup(attr attribute.KeyValue) attribute.KeyValue {
+	depthLimit := r.attrValueDepthLimit()
+	if !r.allowDupKeys {
+		var deduplicated bool
+		if depthLimit < 0 {
+			attr, deduplicated = attrnorm.KeyValue(attr)
+		} else {
+			attr, _, deduplicated = attrnorm.KeyValueWithDepthLimit(attr, depthLimit)
+		}
+		if deduplicated {
+			logKeyValuePairDropped()
+		}
+	} else if depthLimit >= 0 {
+		attr, _ = attrnorm.KeyValueLimitDepth(attr, depthLimit)
+	}
+	return attr
+}
+
+func (r *Record) attrValueDepthLimit() int {
+	if r.attributeValueDepthLimit == 0 {
+		return defaultAttrValDepthLim
+	}
+	return r.attributeValueDepthLimit
 }