+++ sdk/trace/span.go
@@ -278,7 +278,7 @@ func (s *recordingSpan) SetAttributes(attributes ...attribute.KeyValue) {
 			s.addDroppedAttr(1)
 			continue
 		}
-		a = dedupAttr(a)
+		a = dedupAttr(a, s.tracer.provider.spanLimits.AttributeValueDepthLimit)
 		a = attrnorm.Truncate(s.tracer.provider.spanLimits.AttributeValueLengthLimit, a)
 		s.attributes = append(s.attributes, a)
 		s.attributesDirty = true
@@ -353,7 +353,7 @@ func (s *recordingSpan) addOverCapAttrs(limit int, attrs []attribute.KeyValue) {
 
 		if idx, ok := exists[a.Key]; ok {
 			// Perform all updates before dropping, even when at capacity.
-			a = dedupAttr(a)
+			a = dedupAttr(a, s.tracer.provider.spanLimits.AttributeValueDepthLimit)
 			a = attrnorm.Truncate(s.tracer.provider.spanLimits.AttributeValueLengthLimit, a)
 			s.attributes[idx] = a
 			continue
@@ -364,7 +364,7 @@ func (s *recordingSpan) addOverCapAttrs(limit int, attrs []attribute.KeyValue) {
 			// updates are checked and performed.
 			s.addDroppedAttr(1)
 		} else {
-			a = dedupAttr(a)
+			a = dedupAttr(a, s.tracer.provider.spanLimits.AttributeValueDepthLimit)
 			a = attrnorm.Truncate(s.tracer.provider.spanLimits.AttributeValueLengthLimit, a)
 			s.attributes = append(s.attributes, a)
 			exists[a.Key] = len(s.attributes) - 1
@@ -372,10 +372,14 @@ func (s *recordingSpan) addOverCapAttrs(limit int, attrs []attribute.KeyValue) {
 	}
 }
 
-func dedupAttr(attr attribute.KeyValue) attribute.KeyValue {
+func dedupAttr(attr attribute.KeyValue, depthLimit int) attribute.KeyValue {
 	switch attr.Value.Type() {
 	case attribute.SLICE, attribute.MAP:
-		attr, _ = attrnorm.KeyValue(attr)
+		if depthLimit < 0 {
+			attr, _ = attrnorm.KeyValue(attr)
+		} else {
+			attr, _, _ = attrnorm.KeyValueWithDepthLimit(attr, depthLimit)
+		}
 		return attr
 	default:
 		return attr
@@ -538,7 +542,10 @@ func (s *recordingSpan) AddEvent(name string, o ...trace.EventOption) {
 // This method assumes s.mu.Lock is held by the caller.
 func (s *recordingSpan) addEvent(name string, o ...trace.EventOption) {
 	c := trace.NewEventConfig(o...)
-	attrs, _ := attrnorm.KeyValues(c.Attributes())
+	attrs, _, _ := attrnorm.KeyValuesWithDepthLimit(
+		c.Attributes(),
+		s.tracer.provider.spanLimits.AttributeValueDepthLimit,
+	)
 	e := Event{Name: name, Attributes: attrs, Time: c.Timestamp()}
 
 	// Discard attributes over limit.
@@ -718,7 +725,10 @@ func (s *recordingSpan) AddLink(link trace.Link) {
 		return
 	}
 
-	attrs, _ := attrnorm.KeyValues(link.Attributes)
+	attrs, _, _ := attrnorm.KeyValuesWithDepthLimit(
+		link.Attributes,
+		s.tracer.provider.spanLimits.AttributeValueDepthLimit,
+	)
 	l := Link{SpanContext: link.SpanContext, Attributes: attrs}
 
 	// Discard attributes over limit.